refactor: 질문 생성 페이지 리팩토링
