+++ frontend/src/components/questions/create/QuestionInputSection/index.tsx
@@ -1,7 +1,7 @@
 import SelectTitle from "@/components/common/Text/SelectTitle";
 import QuestionInput from "./QuestionInput";
 import QuestionList from "./QuestionList";
-import useQuestionFormStore from "@/stores/useQuestionFormStore";
+import useQuestionFormStore from "@/pages/CreateQuestionPage/stores/useQuestionFormStore";
 
 const QuestionInputSection = () => {
   const questionList = useQuestionFormStore((state) => state.questionList);
