+++ frontend/src/pages/CreateQuestionPage/CreateQuestionPage.tsx
@@ -1,4 +1,4 @@
-import QuestionForm from "@/components/questions/create/QuestionForm";
+import QuestionForm from "@/components/questions/create/CategorySection";
 import { IoArrowBackSharp } from "react-icons/io5";
 import { Link } from "react-router-dom";
 import SidebarPageLayout from "@components/layout/SidebarPageLayout.tsx";
