+++ frontend/src/pages/SessionPage/types/session.d.ts
