+++ frontend/src/pages/SessionPage/view/SessionSidebar.tsx
@@ -2,7 +2,7 @@ import { FaClipboardList, FaFolder } from "react-icons/fa";
 import { FaUserGroup } from "react-icons/fa6";
 import { Socket } from "socket.io-client";
 import { TbCrown } from "react-icons/tb";
-import { Question } from "@/pages/SessionPage/type/session";
+import { Question } from "@/pages/SessionPage/types/session";
 
 interface ParticipantsData {
   nickname: string;
